requirements.txt
accuweather==4.2.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiohttp-retry==2.9.1
aiosignal==1.4.0
alpaca-py==0.39.0
amadeus==12.0.0
annotated-types==0.7.0
anyio==4.9.0
astroid==3.3.11
attrs==25.3.0
audioop-lts==0.2.1
Authlib==0.14.3
backoff==2.2.1
black==25.1.0
blessed==1.20.0
bpython==0.25
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coverage==7.9.2
crcmod==1.7
cryptography==44.0.2
curtsies==0.4.2
cwcwidth==0.1.10
dill==0.4.0
distro==1.9.0
dnspython==2.7.0
easypost==10.1.0
email_validator==2.2.0
enum34==1.1.10
evernote3==1.25.14
execnet==2.1.1
fastapi==0.115.11
fastapi-cli==0.0.7
finnhub-python==2.4.23
firebase-admin==6.7.0
flake8==7.2.0
frozenlist==1.7.0
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-kms==3.5.1
google-cloud-secret-manager==2.23.3
google-cloud-storage==3.1.0
google-crc32c==1.7.0
google-resumable-media==2.7.2
google_search_results==2.4.2
googleapis-common-protos==1.69.2
greenlet==3.1.1
griffe==1.6.1
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.14.0
html2text==2025.4.15
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
iniconfig==2.1.0
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mcp==1.11.0
mdurl==0.1.2
mock==5.2.0
msgpack==1.1.0
multidict==6.6.3
mypy==1.16.1
mypy_extensions==1.1.0
nulltype==2.3.1
numpy==2.2.4
oauth2==1.9.0.post1
oauthlib==3.2.2
openai==1.98.0
openai-agents==0.2.4
openapi-python-client==0.24.3
orjson==3.11.0
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
phonenumbers==9.0.9
plaid-python==31.0.0
platformdirs==4.3.8
pluggy==1.6.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.10.6
pydantic-extra-types==2.10.5
pydantic-settings==2.8.1
pydantic_core==2.27.2
pyflakes==3.3.2
Pygments==2.19.1
PyJWT==2.10.1
pylint==3.3.7
pyparsing==3.2.1
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-cov==6.2.1
pytest-mock==3.14.1
pytest-timeout==2.4.0
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
pyxdg==0.28
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rich-toolkit==0.13.2
rotten_tomatoes_client==0.0.3
rpds-py==0.26.0
rsa==4.9
ruamel.yaml==0.18.12
ruamel.yaml.clib==0.2.12
ruff==0.11.12
scipy==1.16.0
segment-analytics-python==2.3.4
sentry-sdk==2.35.0
serpapi==0.1.5
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
slack_sdk==3.35.0
sniffio==1.3.1
soxr==0.5.0.post1
spotipy==2.25.1
sse-starlette==2.2.1
sseclient-py==1.8.0
starlette==0.46.1
tomlkit==0.13.3
tqdm==4.67.1
tweepy==4.15.0
twilio==9.6.5
typer==0.15.2
types-requests==2.32.0.20250306
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
wcwidth==0.2.13
websockets==15.0.1
yarl==1.20.1
